Add esm support
Hey! I'd like to add es modules support to this module. There are a few
reasons for that

- one step forward for ecosystem where we could get rid from commonjs in
favour of standard
- rollup bundler could build this dependency without commonjs plugin
which converts cjs to esm
- jss-plugin-camel-case project could be used directly in the browser
with the help of unpkg service

```js
import camelCase from 'https://unpkg.com/jss-plugin-camel-case@next?module'
```

I used rollup to convert esm to cjs because it has only a few small
dependencies and is the only clean way to provide `module.exports`
instead of `exports.default`.